fixed more requested issues, we avisorated the router as you wished relates #24
